+++ LethalRichPresence/Core/PlaceholderResolver.cs
@@ -12,20 +12,22 @@ public static class PlaceholderResolver
     Dictionary<string, string> placeholders = new(){
       {"currentplanet", Variables.CurrentPlanet()},
       {"currentweather", Variables.CurrentWeather()},
-      {"quotacountifier", Variables.QuotaNoCountifier()},
       {"quota", Variables.Quota().ToString()},
+      {"quotafullfilled", Variables.QuotaFulfilled().ToString()},
+      {"quotacountifier", Variables.QuotaNoCountifier()},
       {"collected", Variables.LootValue().ToString()},
       {"timeleft", Variables.TimeRemaining().ToString()},
-      {"onlineorlan", Variables.IsOnlineOrLAN().ToString()}
+      {"onlineorlan", Variables.IsOnlineOrLAN().ToString()},
+      {"hosting", Variables.IsHostingOrMember().ToString()},
+      {"partyprivacy", Variables.PartyPrivacy().ToString()},
 
     };
     return placeholders;
   }
 
-  public static string ResolvePlaceholders(string input)
+  public static string ResolvePlaceholders(string input, Dictionary<string, string> placeholders)
   {
     string output = input;
-    Dictionary<string, string> placeholders = PlaceholderDictionary();
 
     // detect %string% placeholders and create switch to replace them
     Regex regex = new Regex(@"\%.+?\%");
@@ -37,12 +39,14 @@ public static class PlaceholderResolver
       bool toLower = false;
       string placeholder = match.Value.Replace("%", "");
 
+      // if there's & in front of the placeholder, make it lowercase and remove spaces
       if (placeholder.StartsWith("&"))
       {
         toLower = true;
         placeholder = placeholder.Replace("&", "");
       }
 
+      // actual resolving placeholders
       if (placeholders.ContainsKey(placeholder))
       {
         output = output.Replace(match.Value, placeholders[placeholder]);
@@ -51,7 +55,8 @@ public static class PlaceholderResolver
       }
     }
 
-    Plugin.logger.LogDebug($"Resolved |{input}| to |{output}|");
+    if (ConfigManager.Debug.Value) Plugin.logger.LogDebug($"Resolved |{input}| to |{output}|");
+
     return output;
   }