+++ Thunderstore/CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.1.1
 
-- fixed a bug where the mod would stop working after starting ship from a planet
+- fixed a bug where the mod would stop working after starting ship from a planet (thanks, [DHStyle541](https://github.com/DHStyle541)!)
 
 # 0.2.0
 
@@ -11,3 +11,24 @@
 - fixed party size/max size being set to null and disabling the mod
 - small optimization in resolving placeholders
 - fixed the option to join lobbies through Discord
+
+# 0.2.1
+
+- added main menu image (thanks, [SheepCommander](https://github.com/SheepCommander)!)
+
+# 0.2.2
+
+- fixed debug messages being logged always
+- optimization: all placeholder values are pulled once per update
+- party size is updated only in orbit
+- added new placeholders: %onlineorlan%, %hosting%, %partyprivacy%
+
+# 0.3.0
+
+### Please re-generate your config file!
+
+- added option to allow for join option to work only in public lobbies (thanks, [1A3Dev](https://github.com/1A3Dev)!)
+
+# 0.3.1
+
+- changed lobby size detection to support MoreCompany/BiggerLobby (thanks, [FLozyXD](https://github.com/FLozyXD)!)